[MOVE/#9] UiState 패키지 이동
