+++ app/src/main/java/org/sopt/sample/data/type/BaseUrlType.kt
@@ -1,4 +1,4 @@
-package org.sopt.sample.data.local
+package org.sopt.sample.data.type
 
 enum class BaseUrlType {
     IN_SOPT, REQRES, MUSIC
