+++ app/src/main/java/org/sopt/sample/presentation/login/LoginActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import androidx.activity.viewModels
 import dagger.hilt.android.AndroidEntryPoint
 import org.sopt.sample.R
-import org.sopt.sample.data.local.UiState
+import org.sopt.sample.util.UiState
 import org.sopt.sample.databinding.ActivityLoginBinding
 import org.sopt.sample.presentation.login.LoginViewModel.Companion.INCORRECT_EMAIL_CODE
 import org.sopt.sample.presentation.login.LoginViewModel.Companion.INCORRECT_PWD_CODE
