+++ app/src/main/java/org/sopt/sample/presentation/main/gallery/GalleryViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import org.sopt.sample.data.dto.request.RequestRegisterMusicDto
-import org.sopt.sample.data.local.UiState
+import org.sopt.sample.util.UiState
 import org.sopt.sample.data.repository.MusicRepository
 import org.sopt.sample.util.ContentUriRequestBody
 import retrofit2.HttpException
