+++ app/src/main/java/org/sopt/sample/presentation/main/home/HomeViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import org.sopt.sample.data.dto.response.ResponseGetFollowerListDto.Follower
-import org.sopt.sample.data.local.UiState
+import org.sopt.sample.util.UiState
 import org.sopt.sample.data.repository.FollowerRepository
 import timber.log.Timber
 import javax.inject.Inject
