+++ app/src/main/java/org/sopt/sample/presentation/main/music/MusicViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import org.sopt.sample.data.dto.response.ResponseMusicDto
-import org.sopt.sample.data.local.UiState
+import org.sopt.sample.util.UiState
 import org.sopt.sample.data.repository.MusicRepository
 import timber.log.Timber
 import javax.inject.Inject
